% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWeatherVariables.R
\name{calculateActualVapourPressure}
\alias{calculateActualVapourPressure}
\title{calculateActualVapourPressure.}
\usage{
calculateActualVapourPressure(rhMin, rhMax, tMin, tMax)
}
\arguments{
\item{-}{rhMin: double}

\item{-}{rhMax: double}

\item{-}{tMin: double}

\item{-}{tMax: double}
}
\value{
double
}
\description{
\code{calculateActualVapourPressure} function to calculate the actual vapour pressure
}
\details{
function to calculate the actual vapour pressure from min/max RH and temp
eqn 15 in FAO 56
}
\examples{
NEED TO INSERT

}
\references{
see aWhere API Documentation (insert HTTP address)
}
