% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWeatherVariables.R
\name{calculateAerodynamicResistance}
\alias{calculateAerodynamicResistance}
\title{calculateAerodynamicResistance.}
\usage{
calculateAerodynamicResistance(windSpeed, cropHeight)
}
\arguments{
\item{-}{windspeed: In m/s measured at 2m - double}

\item{-}{cropHeight: in m}
}
\value{
double
}
\description{
\code{calculateAerodynamicResistance} function to calculate aerodynamic resistance for a given crop
}
\details{
function to calculate aerodynamic resistance for a given crop
takes crop height (m) and wind speed as inputs
EQN 4 in FAO 56

zW height of wind measurements [m],
zH height of humidity measurements [m],
zero plane displacement height [m],
zOM roughness length governing momentum transfer [m],
zOH roughness length governing transfer of heat and vapour [m],
k von Karman's constant, 0.41 [-],
z wind speed at height z [m s-1].
}
\examples{
NEED TO INSERT

}
\references{
see aWhere API Documentation (insert HTTP address)
}
