% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWeatherVariables.R
\name{calculateAtmosphericDensity}
\alias{calculateAtmosphericDensity}
\title{calculateAtmosphericDensity.}
\usage{
calculateAtmosphericDensity(temp, RH, elev)
}
\arguments{
\item{-}{temp: In C - double}

\item{-}{RH: In C - double}

\item{-}{elev: In C - double}
}
\value{
double
}
\description{
\code{calculateAtmosphericDensity} function for calcuating the atmospheric density
}
\details{
function to calculate the atmospheric density from temp and atmospheric pressure
FAO 56 3-5 and 3-6
}
\examples{
NEED TO INSERT

}
\references{
see aWhere API Documentation (insert HTTP address)
}
