% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWeatherVariables.R
\name{calculateAtmosphericPressure}
\alias{calculateAtmosphericPressure}
\title{calculateAtmosphericPressure.}
\usage{
calculateAtmosphericPressure(elevation, elevAtRefLevel, temp)
}
\arguments{
\item{-}{elevation: In m - double}

\item{-}{elevAtRefLevel: In m - double}

\item{-}{temp: In C - double}
}
\value{
double
}
\description{
\code{calculateAtmosphericPressure} #function for calculating Atmospheric Pressure
}
\details{
function for calculating Atmospheric Pressure: equantion (3-2) in FAO 56

P atmospheric pressure at elevation z [kPa]
Po atmospheric pressure at sea level = 101.3 [kPa]
z elevation [m]
zo elevation at reference level [m] (normally 2)
g gravitational acceleration = 9.807 [m s-2]
R specific gas constant = 287 [J kg-1 K-1]
al constant lapse rate moist air = 0.0065 [K m-1]
TKo reference temperature [K] at elevation zo given by
TKo = 273.16 + T (3-3)
where: T mean air temperature for the time period of calculation [oC]
}
\examples{
NEED TO INSERT

}
\references{
see aWhere API Documentation (insert HTTP address)
}
