% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWeatherVariables.R
\name{calculateBoundaryLayerResistance}
\alias{calculateBoundaryLayerResistance}
\title{calculateBoundaryLayerResistance.}
\usage{
calculateBoundaryLayerResistance(dimMockLeaf, wind)
}
\arguments{
\item{-}{dimMockLeaf: double}

\item{-}{wind: in m/s - double}
}
\value{
double
}
\description{
\code{calculateBoundaryLayerResistance} function to calculate boundary layer resistance
}
\details{
function to calculate Boundary Layer Resistance for LWR model info
comes http://agsys.cra-cin.it/tools/leafwetness/help/
}
\examples{
NEED TO INSERT

}
\references{
see aWhere API Documentation (insert HTTP address)
}
