% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWeatherVariables.R
\name{calculateDewPointDepression}
\alias{calculateDewPointDepression}
\title{calculateDewPointDepression.}
\usage{
calculateDewPointDepression(airTemp, RH, temp)
}
\arguments{
\item{-}{airTemp: in C - double}

\item{-}{RH: 1-100 - double}

\item{-}{temp: in C double}
}
\value{
double
}
\description{
\code{calculateDewPointDepression} function to calculate the dew point depression
}
\details{
function to calculate Dew Point Depression
definition from Kim, 2002 paper
}
\examples{
NEED TO INSERT

}
\references{
see aWhere API Documentation (insert HTTP address)
}
