% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWeatherVariables.R
\name{calculateDewPointTemp}
\alias{calculateDewPointTemp}
\title{calculateDewPointTemp.}
\usage{
calculateDewPointTemp(actualVapourPressure)
}
\arguments{
\item{-}{actualVapourPressure: double}
}
\value{
double
}
\description{
\code{calculateDewPointTemp} function for calculating Air dew point temperature
}
\details{
function for calculating Air dew point temperature
Equation 3-11 in FAO 56
}
\examples{
NEED TO INSERT

}
\references{
see aWhere API Documentation (insert HTTP address)
}
