% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWeatherVariables.R
\name{calculateLandsbergWind}
\alias{calculateLandsbergWind}
\title{calculateLandsbergWind.}
\usage{
calculateLandsbergWind(wind, cropHeight, canopyHeight)
}
\arguments{
\item{-}{wind: in m/s - double}

\item{-}{cropHeight: in m - double}

\item{-}{canopyHeight: in m - double}
}
\value{
double
}
\description{
\code{calculateLandsbergWind} function to calculate adjusted windspeed
}
\details{
function to calculate Lansberg's Adjusted windspeed
}
\examples{
NEED TO INSERT

}
\references{
see aWhere API Documentation (insert HTTP address)
}
