% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWeatherVariables.R
\name{calculateLatentHeatVap}
\alias{calculateLatentHeatVap}
\title{calculateLatentHeatVap.}
\usage{
calculateLatentHeatVap()
}
\value{
double
}
\description{
\code{calculateLatentHeatVap} function for calculating the latent heat of vaporization
}
\details{
function for calculating the latent heat of vaporization
FAO 56 simplifies by using a single value over all temps as it changes little
2.45 MJ kg-1.
}
\examples{
NEED TO INSERT

}
\references{
see aWhere API Documentation (insert HTTP address)
}
