% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWeatherVariables.R
\name{calculatePsychrometricConstant}
\alias{calculatePsychrometricConstant}
\title{calculatePsychrometricConstant.}
\usage{
calculatePsychrometricConstant(elevation, elevAtRefLevel, temp)
}
\arguments{
\item{-}{elevation: in m - double}

\item{-}{elevAtRefLevel: in m - double}

\item{-}{temp: in C double}
}
\value{
double
}
\description{
\code{calculatePsychrometricConstant} function to calculate the psychometric contstant
}
\details{
function to calculate the physchometric constant based on on atmospheric
pressure/elevation
eqn 3-10 in FAO 56
}
\examples{
NEED TO INSERT

}
\references{
see aWhere API Documentation (insert HTTP address)
}
