% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWeatherVariables.R
\name{calculateRosenbergWind}
\alias{calculateRosenbergWind}
\title{calculateRosenbergWind.}
\usage{
calculateRosenbergWind(wind, cropHeight)
}
\arguments{
\item{-}{wind: in m/s - double}

\item{-}{cropHeight: in m - double}
}
\value{
double
}
\description{
\code{calculateRosenbergWind} function to calculate adjusted windspeed
}
\details{
function to calculate Rosenberg's Adjusted windspeed for the CART model info
comes http://agsys.cra-cin.it/tools/leafwetness/help/ 
paper is Model to Enhance Site-Specific Estimation of Leaf Wetness Duration by Kim et al, 2002
}
\examples{
NEED TO INSERT

}
\references{
see aWhere API Documentation (insert HTTP address)
}
