% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWeatherVariables.R
\name{calculateRoughnessLength}
\alias{calculateRoughnessLength}
\title{calculateRoughnessLength.}
\usage{
calculateRoughnessLength(cropHeight)
}
\arguments{
\item{-}{cropHeight: In m - double}
}
\value{
double
}
\description{
\code{calculateRoughnessLength} function to calculate the roughness length
}
\details{
function to calculate the roughness length for aerodynamic resistance calculations for a given crop
takes crop height (m).  Note that I am using the values from FAO 56 instead of the CART model paper
EQN 4 in FAO 56
}
\examples{
NEED TO INSERT

}
\references{
see aWhere API Documentation (insert HTTP address)
}
