% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWeatherVariables.R
\name{calculateSaturationVapourPressure}
\alias{calculateSaturationVapourPressure}
\title{calculateSaturationVapourPressure.}
\usage{
calculateSaturationVapourPressure(temp)
}
\arguments{
\item{-}{temp: In C - double}
}
\value{
double
}
\description{
\code{calculateSaturationVapourPressure} function for calcuating the saturation vapour pressure
}
\details{
function to calculate the saturation vapour pressure
FAO 3-8, temp in C
}
\examples{
NEED TO INSERT

}
\references{
see aWhere API Documentation (insert HTTP address)
}
