% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWeatherVariables.R
\name{calculateSlopeOfVapourPressureCurve}
\alias{calculateSlopeOfVapourPressureCurve}
\title{calculateSlopeOfVapourPressureCurve.}
\usage{
calculateSlopeOfVapourPressureCurve(temp)
}
\arguments{
\item{-}{temp: In C - double}
}
\value{
double
}
\description{
\code{calculateSlopeOfVapourPressureCurve} function for calcuating the slope of the vapour pressure curve
}
\details{
function for calcuating the slope of the vapour pressure curve based on temp
temp in C
eqn 13 from FAO 56
}
\examples{
NEED TO INSERT

}
\references{
see aWhere API Documentation (insert HTTP address)
}
