% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWeatherVariables.R
\name{calculateSpecificHeatAir}
\alias{calculateSpecificHeatAir}
\title{calculateSaturationVapourPressure.}
\usage{
calculateSpecificHeatAir()
}
\value{
double
}
\description{
\code{calculateSaturationVapourPressure} function to return the specific heat of moist air
}
\details{
function to return the specific heat of moist air
variable in FAO 56 3-10
}
\examples{
NEED TO INSERT

}
\references{
see aWhere API Documentation (insert HTTP address)
}
