% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWeatherVariables.R
\name{calculateZeroPlaneDisplacement}
\alias{calculateZeroPlaneDisplacement}
\title{calculateZeroPlaneDisplacement.}
\usage{
calculateZeroPlaneDisplacement(cropHeight)
}
\arguments{
\item{-}{cropHeight: In m - double}
}
\value{
double
}
\description{
\code{calculateZeroPlaneDisplacement} function to calculate the zero plane displacement
}
\details{
function to calculate the zero plane displacement  for aerodynamic resistance
calculations for a given crop takes crop height (m).  Values from FAO 56
}
\examples{
NEED TO INSERT

}
\references{
see aWhere API Documentation (insert HTTP address)
}
